refactore transfers modal and add publish dao
