+++ packages/web-app/src/utils/constants.ts
@@ -71,6 +71,13 @@ export const enum TransferTypes {
   Withdraw = 'Withdraw',
 }
 
+export enum TransactionState {
+  WAITING = 'WAITING',
+  LOADING = 'LOADING',
+  SUCCESS = 'SUCCESS',
+  ERROR = 'ERROR',
+}
+
 export const URL_PATTERN =
   /^(http:\/\/www\.|https:\/\/www\.|http:\/\/|https:\/\/)?[a-z0-9]+([-.]{1}[a-z0-9]+)*\.[a-z]{2,5}(:[0-9]{1,5})?(\/.*)?$/i;
